LPS-66685 Move KBUtil to internal package
